Added some styling. Logout works.
