+++ src/Components/css/MainPage.css
@@ -2,3 +2,44 @@
     display: flex;
     flex-direction: row;
 }
+
+.video-list {
+    display: flex;
+    flex-direction: column;
+}
+
+.category-title {
+    display: flex;
+    flex-direction: row;
+}
+
+.video-list {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: space-between;
+}
+
+.add-buttons {
+    flex-basis: 2;
+    display: flex;
+    flex-direction: row;
+    align-content: flex-end;
+}
+
+.video-card {
+    border-radius: 5px;
+    border: solid 1px black;
+    background-color: ghostwhite;
+    margin-bottom: 2em;
+    flex: 0 1 24%;
+}
+
+.video-card-content {
+    padding: .5em;
+}
+
+.video-card-content h4 {
+    margin-top: 0;
+    margin-bottom: .5em;
+    font-weight: bold;
+}
