+++ src/Components/css/Sidebar.css
@@ -1,3 +1,4 @@
 .sidebar {
     padding-left: 1em;
+    padding-right: 1em;
 }
