+++ src/Components/videos/VideoCard.js
@@ -3,8 +3,10 @@ import React from 'react'
 const VideoCard = props => {
     
     return (
-        <div className='videoCard'>
-            <p>{props.video.videoTitle}</p>
+        <div className='video-card'>
+            <div className='video-card-content'>
+                <h4>{props.video.videoTitle}</h4>
+            </div>
         </div>
     )
 }
