+++ src/Components/videos/VideoList.js
@@ -3,11 +3,13 @@ import {createPortal} from 'react-dom'
 import VideoForm from './VideoForm'
 import dataManager from '../../modules/dataManager'
 import VideoDeck from './VideoDeck'
+import CategoryForm from '../categories/CategoryForm'
 
 const VideoList = props => {
 
     const [videos, setVideos] = useState([]);
     const [showVideoForm, setShowVideoForm] = useState(false);
+    const [showCategoryForm, setShowCategoryForm] = useState(false);
 
     const getVideos = () => {
         return dataManager.getByProperty('videos', 'userId', parseInt(sessionStorage.getItem('user')))
@@ -18,6 +20,10 @@ const VideoList = props => {
         showVideoForm ? setShowVideoForm(false) : setShowVideoForm(true)
     }
 
+    const toggleCategoryForm = () => {
+        showCategoryForm ? setShowCategoryForm(false) : setShowCategoryForm(true)
+    }
+
     useEffect(() => {
         getVideos();
     }, [])
@@ -25,7 +31,7 @@ const VideoList = props => {
     const modalDiv = document.getElementById('modal');
 
     return (
-        <section className='videoList'>
+        <section className='video-list'>
             {showVideoForm
                 ? createPortal(<VideoForm
                                     categories={props.categories}
@@ -36,14 +42,27 @@ const VideoList = props => {
                                 />, modalDiv)
                 : null
             }
-            <p onClick={toggleVideoForm}>New video</p>
-            {props.categories.map(mappedCategory =>
-                <VideoDeck
-                    videos={videos}
-                    key={mappedCategory.id}
-                    category={mappedCategory}
-                />
-            )}
+            {showCategoryForm
+                ? createPortal(<CategoryForm
+                                    categories={props.categories}
+                                    getCategories={props.getCategories}
+                                    toggleCategoryForm={toggleCategoryForm}
+                                />, modalDiv)
+                : null
+            }
+            <div className='add-buttons'>
+                <p onClick={toggleVideoForm}> &#x2b; New video</p>
+                <p onClick={toggleCategoryForm}> &#x2b; New category</p>
+            </div>
+            <div className='video-decks'>
+                {props.categories.map(mappedCategory =>
+                    <VideoDeck
+                        videos={videos}
+                        key={mappedCategory.id}
+                        category={mappedCategory}
+                    />
+                )}
+            </div>
         </section>
     )
 }
